+++ .github/workflows/pull-request.yml
@@ -27,6 +27,11 @@ jobs:
       - name: Checkout the Repository
         uses: actions/checkout@v4
 
+      - name: Set up Node.js
+        uses: actions/setup-node@v4
+        with:
+          node-version: '20.x'
+
       - name: Install Dependencies
         run: npm install
 
@@ -49,7 +54,6 @@ jobs:
 
       - name: Run linting check
         if: steps.changed-files.outputs.only_changed != 'true'
-        
         run: npm run lint:check
         
       - name: Check for localStorage Usage
@@ -66,6 +70,7 @@ jobs:
   Check-Changed-Files:
     name: File count, sensitive files and branch check
     runs-on: ubuntu-latest
+    # needs: Code-Quality-Checks
     steps:
       - name: Checkout code
         uses: actions/checkout@v4
@@ -73,19 +78,40 @@ jobs:
       - name: Set up Python
         uses: actions/setup-python@v5
         with:
-          python-version: 3.11
+          python-version: 3.9
 
       - name: Run Python script
         run: |
-          python3 .github/workflows/check_files.py --base_branch "${{ github.base_ref }}" --pr_branch "${{ github.head_ref }}" --sensitive_files .github/workflows/sensitive_files.txt --verbose
+          python .github/workflows/count_changed_files.py --base_branch "${{ github.base_ref }}" --pr_branch "${{ github.head_ref }}"
+
+  Check-ESlint-Disable:
+    name: Check for eslint-disable
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+
+      - name: Set up Python
+        uses: actions/setup-python@v5
+        with:
+          python-version: 3.9
+
+      - name: Run Python script
+        run: |
+          python .github/workflows/eslint_disable_check.py
 
   Test-Application:
     name: Test Application
     runs-on: ubuntu-latest
-    needs: [Code-Quality-Checks]
+    needs: [Code-Quality-Checks, Check-ESlint-Disable]
     steps:
       - name: Checkout the Repository
         uses: actions/checkout@v4
+
+      - name: Set up Node.js
+        uses: actions/setup-node@v4
+        with:
+          node-version: '20.x'
 
       - name: Install Dependencies
         run: npm install
